+++ src/components/bug-report/index.js
@@ -1,3 +1,3 @@
 import "./bug-report.css";
 
-export { bugReportMethods } from "./bug-report.js";
+export { bugReportMethods, BUG_REPORT_RICKROLL_EMBED_URL, BUG_REPORT_RICKROLL_URL } from "./bug-report.js";